agent-simple/pyproject.toml
[tool.poetry]
name = "agent-simple"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "~=0.103"
uvicorn = "~=0.23"
langchain = "~=0.0"
langchain-community = "~=0.0"
openai = "0.28"
python-dotenv = "~=1.0"
tiktoken = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
